Main android files for the Project
This are the only files required to make the program. the mipmap-nodpi folder is not made by default when making an android project so you should generate it on your own. Reason why we put our Image in the mipmap-nodpi folder is to make sure that the image doesn't resize according to the device screen size thous making the coordinates useless.